refactor: project structure reorganization (from all functions in a single file to a grouping by context)
